# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Torneo Socorro Senior Master Plus 40 - Tournament Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /arbitros:
    get:
      operationId: getArbitros
      tags: [arbitros]
      summary: List referees (roster)
      parameters:
        - name: activo
          in: query
          required: false
          schema:
            type: boolean
        - name: buscar
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Referees, most active (by matches worked) first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Arbitro"
    post:
      operationId: crearArbitro
      tags: [arbitros]
      summary: Register a new referee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArbitroInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Arbitro"
...
  /arbitros/{id}:
    patch:
      operationId: actualizarArbitro
      tags: [arbitros]
      summary: Edit a referee, or deactivate them
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArbitroInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Arbitro"
        "404":
          description: Not found
    delete:
      operationId: borrarArbitro
      tags: [arbitros]
      summary: Delete a referee who has never worked a match
      description: >
        Solo se puede borrar a un árbitro que no tenga ningún partido a su
        nombre — sirve para deshacer un alta equivocada o un duplicado. Si ya
        dirigió partidos se responde 409: ahí lo que corresponde es
        desactivarlo, para no dejar esos partidos sin saber quién los dirigió.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
        "409":
          description: The referee already has matches to their name
  /arbitros/{id}/estadisticas:
    get:
      operationId: getFichaArbitro
      tags: [arbitros]
      summary: Full referee profile — totals, breakdowns, full match history, upcoming assignments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Referee profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArbitroFicha"
        "404":
          description: Not found

  # ── PARTIDOS ───────────────────────────────────────────────────────────────
...
  /egresos:
    get:
      operationId: getEgresos
      tags: [egresos]
      summary: List organization expenses
      parameters:
        - name: categoria
          in: query
          required: false
          schema:
            type: string
        - name: desde
          in: query
          required: false
          description: Fecha mínima (inclusive), para ver un periodo puntual.
          schema:
            type: string
        - name: hasta
          in: query
          required: false
          description: Fecha máxima (inclusive).
          schema:
            type: string
        - $ref: "#/components/parameters/Temporada"
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Egreso"
    post:
      operationId: createEgreso
      tags: [egresos]
      summary: Register an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EgresoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Egreso"
  /egresos/{id}:
    patch:
      operationId: updateEgreso
      tags: [egresos]
      summary: Correct an expense already recorded
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EgresoUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Egreso"
        "404":
          description: Not found
    delete:
      operationId: deleteEgreso
      tags: [egresos]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
components:
  parameters:
...
    Temporada:
      name: temporada
      in: query
      required: false
      schema:
        type: string
  schemas:
...
    Arbitro:
      type: object
      required: [id, nombre, activo, createdAt]
      properties:
        id:
          type: number
        nombre:
          type: string
        telefono:
          type: ["string", "null"]
        # Data URL (base64) de la foto, igual que la del carné del jugador.
        foto:
          type: ["string", "null"]
        activo:
          type: boolean
        notas:
          type: ["string", "null"]
        # De solo lectura: cuántos partidos lleva dirigidos en el torneo que
        # se está consultando. Se manda ya calculado para que el listado no
        # tenga que pedir la ficha de cada uno solo para mostrar esto.
        partidosDirigidos:
          type: number
        createdAt:
          type: string
    ArbitroInput:
      type: object
      required: [nombre]
      properties:
        nombre:
          type: string
        telefono:
          type: string
        # null borra la foto que tuviera; omitirlo la deja como está.
        foto:
          type: ["string", "null"]
        activo:
          type: boolean
        notas:
          type: string
    ArbitroEquipo:
      type: object
      required: [equipoId, equipoNombre, partidos]
      properties:
        equipoId:
          type: number
        equipoNombre:
          type: string
        partidos:
          type: number
    ArbitroFecha:
      type: object
      required: [fecha, partidos]
      properties:
        fecha:
          type: string
        partidos:
          type: number
    ArbitroFase:
      type: object
      required: [fase, partidos]
      properties:
        fase:
          type: string
        partidos:
          type: number
    ArbitroPartidoHistorial:
      type: object
      required: [partidoId, semana, localNombre, visitanteNombre, amarillas, rojas]
      properties:
        partidoId:
          type: number
        fecha:
          type: ["string", "null"]
        semana:
          type: number
        fase:
          type: ["string", "null"]
        localNombre:
          type: string
        visitanteNombre:
          type: string
        golesLocal:
          type: ["number", "null"]
        golesVisitante:
          type: ["number", "null"]
        amarillas:
          type: number
        rojas:
          type: number
        pagado:
          type: ["number", "null"]
    ArbitroProximoPartido:
      type: object
      required: [partidoId, semana, localNombre, visitanteNombre]
      properties:
        partidoId:
          type: number
        fecha:
          type: ["string", "null"]
        hora:
          type: ["string", "null"]
        semana:
          type: number
        fase:
          type: ["string", "null"]
        localNombre:
          type: string
        visitanteNombre:
          type: string
    ArbitroResumen:
      type: object
      required: [partidosDirigidos, amarillas, rojas, totalPagado]
      properties:
        partidosDirigidos:
          type: number
        amarillas:
          type: number
        rojas:
          type: number
        totalPagado:
          type: number
    ArbitroFicha:
      type: object
      required: [arbitro, resumen, porEquipo, porFase, partidosPorFecha, historial, proximos]
      properties:
        arbitro:
          $ref: "#/components/schemas/Arbitro"
        resumen:
          $ref: "#/components/schemas/ArbitroResumen"
        porEquipo:
          type: array
          items:
            $ref: "#/components/schemas/ArbitroEquipo"
        porFase:
          type: array
          items:
            $ref: "#/components/schemas/ArbitroFase"
        partidosPorFecha:
          type: array
          items:
            $ref: "#/components/schemas/ArbitroFecha"
        historial:
          type: array
          items:
            $ref: "#/components/schemas/ArbitroPartidoHistorial"
        proximos:
          type: array
          items:
            $ref: "#/components/schemas/ArbitroProximoPartido"
...
    Egreso:
      type: object
      required: [id, fecha, descripcion, valor]
      properties:
        id:
          type: number
        fecha:
          type: string
        descripcion:
          type: string
        categoria:
          type: ["string", "null"]
        valor:
          type: number
        createdAt:
          type: string
    EgresoInput:
      type: object
      required: [fecha, descripcion, valor]
      properties:
        fecha:
          type: string
          minLength: 1
        descripcion:
          type: string
          minLength: 1
        categoria:
          type: string
        valor:
          type: number
          minimum: 0

    # Para corregir un gasto ya registrado: todos los campos son opcionales,
    # solo se cambia lo que venga.
    EgresoUpdate:
      type: object
      properties:
        fecha:
          type: string
          minLength: 1
        descripcion:
          type: string
          minLength: 1
        categoria:
          type: ["string", "null"]
        valor:
          type: number
          minimum: 0